+++ app/scripts/views/dialog.js
@@ -41,6 +41,8 @@ define([
 
     // QBChat handlers
     chatCallbacksInit: function() {
+      var self = this;
+
       var ContactListView = this.app.views.ContactList,
           MessageView = this.app.views.Message,
           VideoChat = this.app.models.VideoChat,
@@ -73,11 +75,20 @@ define([
         if ('div.popups.is-overlay') {
           $('.is-overlay:not(.chat-occupants-wrap)').removeClass('is-overlay');
         }
-        $('.j-disconnect').addClass('is-overlay');
+        $('.j-disconnect').addClass('is-overlay')
+         .parent('.j-overlay').addClass('is-overlay');
       };
 
-      QB.chat.onReconnectListener = function() {
-        $('.j-disconnect').removeClass('is-overlay');
+      QB.chat.onReconnectListener = function(error) {
+        if (error) {
+          Helpers.log('Error: ', error);
+          self.app.service.checkSession(function (res) {
+            Helpers.log('Check session: ', res);
+          }, 'reconnected');
+        }
+
+        $('.j-disconnect').removeClass('is-overlay')
+         .parent('.j-overlay').removeClass('is-overlay');
       };
 
       currentUser = new Person(_.clone(User.contact), {
@@ -427,18 +438,21 @@ define([
 
         html += '<section class="l-chat-content scrollbar_message"></section>';
         html += '<footer class="l-chat-footer">';
-        html += '<button class="j-toBottom btn_to_bottom"></button>';
-        html += '<div class="l-typing"></div>';
+        html += '<div class="footer_btn j-toBottom btn_to_bottom"></div>';
         html += '<form class="l-message" action="#">';
         html += '<div class="form-input-message textarea" contenteditable="true" placeholder="Type a message"></div>';
-        html += '<button class="btn_message btn_message_smile"><img src="images/icon-smile.svg" alt="smile"></button>';
-        html += '<button class="btn_message btn_message_attach"><img src="images/icon-attach.svg" alt="attach"></button>';
-        html += '<input class="attachment" type="file" accept="image/*">';
-        html += '</form></footer>';
+        html += '<div class="footer_btn j-send_location btn_sendlocation'+((localStorage['QM.latitude'] && localStorage['QM.longitude']) ? ' btn_active' : '')+'"';
+        html += 'data-balloon-length="small" data-balloon="Send you location with messages" data-balloon-pos="up"></div>';
+        html += '<input class="attachment" type="file" accept="image/*"></form>';
+        html += '<div class="l-typing"></div><div class="l-input-menu">';
+        html += '<div class="footer_btn l-input-buttons btn_input_smile j-btn_input_smile" data-balloon="Add smiles" data-balloon-pos="up"></div>';
+        html += '<div class="footer_btn l-input-buttons btn_input_location j-btn_input_location" data-balloon="Send location" data-balloon-pos="up"></div>';
+        html += '<div class="footer_btn l-input-buttons btn_input_attach j-btn_input_attach" data-balloon="Send attachment file" data-balloon-pos="up"></div>';
+        html += '<button class="footer_btn l-input-buttons btn_input_send j-btn_input_send" data-balloon="Send message" data-balloon-pos="up">SEND</button></div></footer>';
 
         html += '</section>';
 
-        $('.l-workspace-wrap .l-workspace').addClass('is-hidden').parent().append(html);
+        $('.l-workspace-wrap .l-workspace').addClass('is-hidden').parent().append($(html));
         textAreaScrollbar();
 
         if (dialog.type === 3 && (!status || status.subscription === 'none'))
@@ -495,7 +509,7 @@ define([
 
       objDom.mCustomScrollbar({
         theme: 'minimal-dark',
-        scrollInertia: 300,
+        scrollInertia: 100,
         mouseWheel: {
           scrollAmount: QMCONFIG.isMac || 'auto',
           deltaFactor: 'auto'
@@ -526,10 +540,6 @@ define([
         live: true
       });
 
-      $(document).on('click', '.j-toBottom', function() {
-        objDom.mCustomScrollbar('scrollTo', 'bottom');
-        $('.j-toBottom').hide();
-      });
     },
 
     createGroupChat: function(type, dialog_id) {
@@ -592,8 +602,12 @@ define([
         isSectionEmpty(list);
 
         // delete chat section
-        if (chat.length > 0) chat.remove();
-        $('#capBox').removeClass('is-hidden');
+        if (chat.is(':visible')) {
+          $('#capBox').removeClass('is-hidden');
+        }
+        if (chat.length > 0) {
+          chat.remove();
+        }
         delete dialogs[dialog_id];
       });