+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 See [the Releases section](https://github.com/QuickBlox/q-municate-web/releases) for changelogs for each release version of Q-municate.
 
+### v. 1.4.0 - May 13, 2016
+* Added ability to send messages whith geotag
+* Added ability to send map with location
+* Added tooltips for footer's buttons
+* Fixed bugs
+
 ### v. 1.3.0 - April 20, 2016
 * Updated QuickBlox JS SDK to 2.1.1
 * Audio/video calls restored (WebRTC)
@@ -13,6 +19,18 @@ See [the Releases section](https://github.com/QuickBlox/q-municate-web/releases)
 ### v. 1.2.1 - March 1, 2016
 * Added local notifications for chat
 * Closed calls (Not working with QuickBlox JS SDK 2.0.3)
+* Fixed bugs
+
+### v. 1.2.0 - January 29, 2016
+* Update QuickBlox JS SDK to 2.0.2
+* Added system notification for chat
+* Added typing statuses
+* Added read/delivered statuses
+* Fixed bugs
+
+### v. 1.1.0 - December 15, 2016
+* Update all dependencies
+* Update QuickBlox JS SDK to 1.17.1
 * Fixed bugs
 
 ### v. 1.2.0 - January 29, 2016
@@ -30,7 +48,7 @@ See [the Releases section](https://github.com/QuickBlox/q-municate-web/releases)
 ### v. 1.0.0 – February 12, 2015
 * [Full screen mode for audio and video calls](#18-full-screen-mode-for-audio-and-video-calls)
 * Improvements of calls
-* Bugs  fixes of the chat session expiration
+* Bugs fixes of the chat session expiration
 
 ### v. 0.9 – January 19, 2014
 * [My profile page](#11-my-profile-page)